refactor: make scroll effect a composable and abstract logic from components
